requirements.txt
Django
openai
python-dotenv
pymodulegen
djangorestframework
markdown
django-filter
psycopg